# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Voicecoach API",
    "summary": "Professor de ingl\u00eas por conversa de \u00e1udio.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/sessions/{session_id}/turns": {
      "post": {
        "tags": [
          "turns"
        ],
        "summary": "Envia a fala do aluno e enfileira o turno",
        "description": "``202`` com o ``turn_id``. O trabalho come\u00e7a depois da resposta.\n\nO cabe\u00e7alho \u00e9 **obrigat\u00f3rio** de prop\u00f3sito. Torn\u00e1-lo opcional (\"se faltar, eu\ngero uma\") faria o esquecimento do cliente virar um turno extra processado e\npago, em sil\u00eancio \u2014 e a rede m\u00f3vel, que \u00e9 o caso de uso inteiro da\nidempot\u00eancia, \u00e9 justamente onde o reenvio acontece.\n\nO rate limit (ADR-0063) roda como depend\u00eancia da ROTA, antes at\u00e9 deste\ncorpo come\u00e7ar \u2014 \u00e9 o que faz a checagem custar zero IO de upload quando\nnega. A cota di\u00e1ria e o kill switch, ao contr\u00e1rio, precisam do\n``student_id`` (s\u00f3 a sess\u00e3o o revela) e por isso moram dentro do\n``handler.handle``.",
        "operationId": "criar_turn_v1_sessions__session_id__turns_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 255,
              "description": "Chave gerada pelo cliente. Reenviar a mesma devolve o mesmo turn_id, sem criar nem reprocessar nada.",
              "title": "Idempotency-Key"
            },
            "description": "Chave gerada pelo cliente. Reenviar a mesma devolve o mesmo turn_id, sem criar nem reprocessar nada."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_criar_turn_v1_sessions__session_id__turns_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/turns/{turn_id}": {
      "get": {
        "tags": [
          "turns"
        ],
        "summary": "O turno completo \u2014 o contrato de recuo (ADR-0026, item 4)",
        "description": "Tudo que o app precisa mostrar, com as URLs j\u00e1 assinadas.\n\nAssinar N trechos s\u00e3o N HMACs locais (microssegundos cada, ADR-0024) \u2014 \u00e9 o\nque torna aceit\u00e1vel entregar as URLs prontas em vez de o cliente pedir uma\npor trecho, que era o roundtrip por frase que o ADR-0024 recusou.",
        "operationId": "obter_turn_v1_turns__turn_id__get",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Turn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/turns/{turn_id}/discard": {
      "post": {
        "tags": [
          "turns"
        ],
        "summary": "'Descartar': o turn some da tela ativa, sem apagar nada (CARD-032)",
        "description": "Idempotente (RNF1): descartar duas vezes \u00e9 o mesmo que uma, `204` as duas.\n\nNada \u00e9 apagado (RF1/RF4) \u2014 o turn continua no hist\u00f3rico e nas agrega\u00e7\u00f5es\n(RNF4/RNF5). O \u00fanico efeito \u00e9 ``discarded_at`` marcado, para o cliente\nparar de mostr\u00e1-lo como turn ativo.",
        "operationId": "descartar_turn_v1_turns__turn_id__discard_post",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Turn Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/turns/{turn_id}/translations": {
      "post": {
        "tags": [
          "turns"
        ],
        "summary": "Traduz para portugu\u00eas um texto do turno (CARD-036)",
        "description": "O aluno pediu a tradu\u00e7\u00e3o de um texto que o produto j\u00e1 produziu.\n\n**O corpo diz QUAL texto, nunca o texto** (RF1). Traduzir de novo o mesmo\ntexto n\u00e3o cobra de novo (RF4) e responde `cached: true`.",
        "operationId": "traduzir_texto_do_turn_v1_turns__turn_id__translations_post",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Turn Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/turns/{turn_id}/events": {
      "get": {
        "tags": [
          "turns"
        ],
        "summary": "Entrega progressiva por SSE (ADR-0026)",
        "description": "O stream. Fecha em ``completed``/``failed``, no prazo, ou no disconnect.\n\n**Quem fecha no disconnect \u00e9 o `sse-starlette`**, e \u00e9 o principal motivo de a\nbiblioteca existir aqui: ele observa a mensagem ASGI ``http.disconnect`` e\n**cancela a corrotina** deste gerador. O cancelamento propaga para o\n``async with`` da assinatura dentro do caso de uso, que desfaz o\n``SUBSCRIBE`` e devolve a conex\u00e3o do Redis. Sem isso, um aluno que fecha o\napp deixaria uma conex\u00e3o pendurada at\u00e9 o prazo de 60 s \u2014 multiplicado por\ntodo mundo.\n\nO `Last-Event-ID` chega por cabe\u00e7alho porque \u00e9 assim que o `EventSource`\nnativo reconecta: ele o manda sozinho, sem o cliente escrever uma linha.",
        "operationId": "acompanhar_turn_v1_turns__turn_id__events_get",
        "parameters": [
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Turn Id"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "O `id:` do \u00faltimo evento recebido. O servidor reenvia s\u00f3 o que vem depois dele, lendo do banco (ADR-0026, item 3).",
              "title": "Last-Event-Id"
            },
            "description": "O `id:` do \u00faltimo evento recebido. O servidor reenvia s\u00f3 o que vem depois dele, lendo do banco (ADR-0026, item 3)."
          }
        ],
        "responses": {
          "200": {
            "description": "Fluxo `text/event-stream`. Cada evento carrega, em `data`, o payload do seu `event:` \u2014 o mapa \u00e9 `TurnEventPayloads`.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TurnEventPayloads"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_criar_turn_v1_sessions__session_id__turns_post": {
        "properties": {
          "audio": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio",
            "description": "A fala do aluno."
          }
        },
        "type": "object",
        "required": [
          "audio"
        ],
        "title": "Body_criar_turn_v1_sessions__session_id__turns_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TranslationRequest": {
        "properties": {
          "target": {
            "$ref": "#/components/schemas/TranslationTarget",
            "description": "`reply` \u00e9 a resposta do professor; `correction` \u00e9 a explica\u00e7\u00e3o de uma corre\u00e7\u00e3o."
          },
          "index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Index",
            "description": "Ignorado para `reply` (h\u00e1 uma resposta por turn); para `correction`, \u00e9 o \u00edndice 0-based na lista de corre\u00e7\u00f5es do turn.",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "target"
        ],
        "title": "TranslationRequest",
        "description": "Qual texto do turn traduzir."
      },
      "TranslationResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "`true` quando a tradu\u00e7\u00e3o j\u00e1 existia e nada foi cobrado. O corpo \u00e9 o mesmo nos dois casos \u2014 o campo torna o 'n\u00e3o paga duas vezes' observ\u00e1vel de fora."
          }
        },
        "type": "object",
        "required": [
          "text",
          "cached"
        ],
        "title": "TranslationResponse",
        "description": "O texto em portugu\u00eas."
      },
...
      "TurnAcceptedResponse": {
        "properties": {
          "turn_id": {
            "type": "string",
            "format": "uuid",
            "title": "Turn Id"
          },
          "replayed": {
            "type": "boolean",
            "title": "Replayed",
            "description": "true quando a Idempotency-Key j\u00e1 tinha sido usada: nenhum turn novo foi criado e nada \u00e9 reprocessado."
          }
        },
        "type": "object",
        "required": [
          "turn_id",
          "replayed"
        ],
        "title": "TurnAcceptedResponse",
        "description": "``202`` do ``POST`` \u2014 o turn est\u00e1 aceito e a caminho."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}